requirements.txt
argcomplete==1.10.3
beautifulsoup4==4.8.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
compressed-rtf==1.0.6
cryptography==38.0.1
customtkinter==4.6.3
darkdetect==0.7.1
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
et-xmlfile==1.1.0
extract-msg==0.28.7
idna==3.4
IMAPClient==2.1.0
lxml==4.9.1
numpy==1.23.3
olefile==0.46
openpyxl==3.0.10
pandas==1.5.0
pdfminer.six==20191110
pdfplumber==0.7.4
Pillow==9.2.0
pycparser==2.21
pycryptodome==3.15.0
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pyxlsb==1.0.9
requests==2.28.1
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
Wand==0.6.10
xlrd==1.2.0
XlsxWriter==3.0.3
yarg==0.1.9